messaging package and rm unecessary packages
