+++ packages/huddle01-interfaces/src/AudioSpacesInterface.ts
@@ -1,16 +1,17 @@
 import { IPeer } from '@huddle01/web-core';
-import { TRoomControls } from './SpacesInterface';
 
-export default abstract class MeetInterface {
+export type TRoomControls = 'muteEveryone';
+
+export default abstract class AudioSpacesInterface {
   abstract joinLobby(spaceId: string): void;
 
   abstract leaveLobby(): void;
 
-  abstract joinMeet(): void;
+  abstract joinSpace(): void;
 
-  abstract leaveMeet(): void;
+  abstract leaveSpace(): void;
 
-  abstract endMeet(): void;
+  abstract endSpace(): void;
 
   abstract fetchAudioStream(deviceId?: string): Promise<MediaStream>;
 
@@ -25,20 +26,6 @@ export default abstract class MeetInterface {
   abstract createMicConsumer(peerId: string): void;
 
   abstract closeMicConsumer(peerId: string): void;
-
-  abstract fetchVideoStream(deviceId?: string): Promise<MediaStream>;
-
-  abstract stopVideoStream(): void;
-
-  abstract produceVideo(camStream: MediaStream, peerIds?: string[]): void;
-
-  abstract stopProducingVideo(): void;
-
-  abstract enumerateCamDevices(): Promise<MediaDeviceInfo[]>;
-
-  abstract createCamConsumer(peerId: string): void;
-
-  abstract closeCamConsumer(peerId: string): void;
 
   abstract getPeer(peerId: string): IPeer;
 
@@ -58,6 +45,10 @@ export default abstract class MeetInterface {
   abstract setAvatar(avatarUrl: string): void;
 
   abstract sendData(peerIds: string[] | '*', data: unknown): void;
+
+  abstract createSpeaker(peerId: string): void;
+
+  abstract createListener(peerId: string): void;
 
   abstract createCohost(peerId: string): void;
