+++ packages/messaging-huddle01/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@dapp-sdk/videomeet-huddle01",
-  "description": "VideoMeet Huddle01 SDK",
+  "name": "@dapp-sdk/messaging-huddle01",
+  "description": "Audio Video Module by Huddle01",
   "version": "0.0.0",
   "main": "dist/index.js",
   "module": "dist/index.js",
@@ -11,7 +11,7 @@
   },
   "license": "MIT",
   "keywords": [
-    "videomeet",
+    "audiospaces",
     "dapp",
     "sdk",
     "interface",
