+++ packages/messaging-huddle01/src/index.ts
@@ -1,0 +1,4 @@
+import AudioSpacesClient from './AudioSpacesClient';
+import VideoMeetClient from './VideoMeetClient';
+
+export { AudioSpacesClient, VideoMeetClient };
