--- packages/videomeet-huddle01/README.md
@@ -1,1 +1,0 @@
-# videomeet-huddle01
